launch/main: split off Launcher and Service into separate files
